refactor: CollectionViewSection으로 분리
